refactor(pk): move walkDirectories and matchPattern back to pk
These functions are internal plan-building utilities only used by
pk/plan.go. Exporting them in pk/repopath polluted the public API
surface with symbols irrelevant to the package's purpose of path
resolution.
